🦁 18th Commit
Added new authors and tags in the admin panel.
Overrode the str method in the Author(while also creating a full_name helper function) and Tag models for better representation.
Created new posts and associated them with authors and tags.
Prepared for enhancements in the admin panel for better post management.